+++ datanode/sqlstore/liquidity_provision.go
@@ -24,8 +24,8 @@ import (
 )
 
 var lpOrdering = TableOrdering{
-	ColumnOrdering{"vega_time", ASC},
-	ColumnOrdering{"id", ASC},
+	ColumnOrdering{Name: "vega_time", Sorting: ASC},
+	ColumnOrdering{Name: "id", Sorting: ASC},
 }
 
 type LiquidityProvision struct {