+++ datanode/sqlstore/orders_test.go
@@ -508,6 +508,10 @@ func TestOrders_CursorPagination(t *testing.T) {
 	t.Run("Should return the page of current orders for a given party when a last and before cursor is given - Newest First", testOrdersCursorPaginationByPartyLastAndBeforeCursorNewestFirst)
 	t.Run("Should return the page of order versions of a given order ID when a last and before cursor is given - Newest First", testOrdersCursorPaginationByOrderIDLastAndBeforeCursorNewestFirst)
 	t.Run("Should return the page of current orders for a given party and market when a last and before cursor is given - Newest First", testOrdersCursorPaginationByMarketAndPartyLastAndBeforeCursorNewestFirst)
+
+	t.Run("Should return all current orders between dates for a given market when no cursor is given", testOrdersCursorPaginationBetweenDatesByMarketNoCursor)
+	t.Run("Should return the first page of current orders between dates for a given market when a first cursor is given", testOrdersCursorPaginationBetweenDatesByMarketFirstCursor)
+	t.Run("Should return the page of current orders for a given market when a first and after cursor is given", testOrdersCursorPaginationBetweenDatesByMarketFirstAndAfterCursor)
 }
 
 type orderTestStores struct {
@@ -845,7 +849,7 @@ func testOrdersCursorPaginationByMarketNoCursor(t *testing.T) {
 	require.NoError(t, err)
 
 	marketID := testData.markets[0].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination, entities.DateRange{})
 
 	require.NoError(t, err)
 	assert.Len(t, got, 5)
@@ -872,7 +876,7 @@ func testOrdersCursorPaginationByPartyNoCursor(t *testing.T) {
 	require.NoError(t, err)
 
 	partyID := testData.parties[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 6)
 	want := append([]entities.Order{}, testData.orders[3], testData.orders[5], testData.orders[7], testData.orders[9], testData.orders[12], testData.orders[13])
@@ -925,7 +929,7 @@ func testOrdersCursorPaginationByMarketAndPartyNoCursor(t *testing.T) {
 
 	partyID := testData.parties[1].ID.String()
 	marketID := testData.markets[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{}, testData.orders[5], testData.orders[9], testData.orders[12])
@@ -951,7 +955,7 @@ func testOrdersCursorPaginationByMarketNoCursorNewestFirst(t *testing.T) {
 	require.NoError(t, err)
 
 	marketID := testData.markets[0].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 5)
 	want := append([]entities.Order{},
@@ -983,7 +987,7 @@ func testOrdersCursorPaginationByPartyNoCursorNewestFirst(t *testing.T) {
 	require.NoError(t, err)
 
 	partyID := testData.parties[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 6)
 	want := append([]entities.Order{},
@@ -1050,7 +1054,7 @@ func testOrdersCursorPaginationByMarketAndPartyNoCursorNewestFirst(t *testing.T)
 
 	partyID := testData.parties[1].ID.String()
 	marketID := testData.markets[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{},
@@ -1083,7 +1087,7 @@ func testOrdersCursorPaginationByMarketFirstCursor(t *testing.T) {
 	require.NoError(t, err)
 
 	marketID := testData.markets[0].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{}, testData.orders[0], testData.orders[3], testData.orders[7])
@@ -1110,7 +1114,7 @@ func testOrdersCursorPaginationByPartyFirstCursor(t *testing.T) {
 	require.NoError(t, err)
 
 	partyID := testData.parties[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{}, testData.orders[3], testData.orders[5], testData.orders[7])
@@ -1165,7 +1169,7 @@ func testOrdersCursorPaginationByMarketAndPartyFirstCursor(t *testing.T) {
 
 	partyID := testData.parties[1].ID.String()
 	marketID := testData.markets[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 2)
 	want := append([]entities.Order{}, testData.orders[5], testData.orders[9])
@@ -1192,7 +1196,7 @@ func testOrdersCursorPaginationByMarketFirstCursorNewestFirst(t *testing.T) {
 	require.NoError(t, err)
 
 	marketID := testData.markets[0].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{},
@@ -1223,7 +1227,7 @@ func testOrdersCursorPaginationByPartyFirstCursorNewestFirst(t *testing.T) {
 	require.NoError(t, err)
 
 	partyID := testData.parties[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{},
@@ -1286,7 +1290,7 @@ func testOrdersCursorPaginationByMarketAndPartyFirstCursorNewestFirst(t *testing
 
 	partyID := testData.parties[1].ID.String()
 	marketID := testData.markets[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 2)
 	want := append([]entities.Order{},
@@ -1318,7 +1322,7 @@ func testOrdersCursorPaginationByMarketLastCursor(t *testing.T) {
 	require.NoError(t, err)
 
 	marketID := testData.markets[0].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{}, testData.orders[7], testData.orders[13], testData.orders[14])
@@ -1345,7 +1349,7 @@ func testOrdersCursorPaginationByPartyLastCursor(t *testing.T) {
 	require.NoError(t, err)
 
 	partyID := testData.parties[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{}, testData.orders[9], testData.orders[12], testData.orders[13])
@@ -1400,7 +1404,7 @@ func testOrdersCursorPaginationByMarketAndPartyLastCursor(t *testing.T) {
 
 	partyID := testData.parties[1].ID.String()
 	marketID := testData.markets[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 2)
 	want := append([]entities.Order{}, testData.orders[9], testData.orders[12])
@@ -1427,7 +1431,7 @@ func testOrdersCursorPaginationByMarketLastCursorNewestFirst(t *testing.T) {
 	require.NoError(t, err)
 
 	marketID := testData.markets[0].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{},
@@ -1458,7 +1462,7 @@ func testOrdersCursorPaginationByPartyLastCursorNewestFirst(t *testing.T) {
 	require.NoError(t, err)
 
 	partyID := testData.parties[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{},
@@ -1521,7 +1525,7 @@ func testOrdersCursorPaginationByMarketAndPartyLastCursorNewestFirst(t *testing.
 
 	partyID := testData.parties[1].ID.String()
 	marketID := testData.markets[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 2)
 	want := append([]entities.Order{},
@@ -1554,7 +1558,7 @@ func testOrdersCursorPaginationByMarketFirstAndAfterCursor(t *testing.T) {
 	require.NoError(t, err)
 
 	marketID := testData.markets[0].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{}, testData.orders[3], testData.orders[7], testData.orders[13])
@@ -1582,7 +1586,7 @@ func testOrdersCursorPaginationByPartyFirstAndAfterCursor(t *testing.T) {
 	require.NoError(t, err)
 
 	partyID := testData.parties[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{}, testData.orders[7], testData.orders[9], testData.orders[12])
@@ -1639,7 +1643,7 @@ func testOrdersCursorPaginationByMarketAndPartyFirstAndAfterCursor(t *testing.T)
 
 	partyID := testData.parties[1].ID.String()
 	marketID := testData.markets[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 1)
 	want := append([]entities.Order{}, testData.orders[9])
@@ -1667,7 +1671,7 @@ func testOrdersCursorPaginationByMarketFirstAndAfterCursorNewestFirst(t *testing
 	require.NoError(t, err)
 
 	marketID := testData.markets[0].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{},
@@ -1699,7 +1703,7 @@ func testOrdersCursorPaginationByPartyFirstAndAfterCursorNewestFirst(t *testing.
 	require.NoError(t, err)
 
 	partyID := testData.parties[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{},
@@ -1764,7 +1768,7 @@ func testOrdersCursorPaginationByMarketAndPartyFirstAndAfterCursorNewestFirst(t 
 
 	partyID := testData.parties[1].ID.String()
 	marketID := testData.markets[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 1)
 	want := append([]entities.Order{},
@@ -1796,7 +1800,7 @@ func testOrdersCursorPaginationByMarketLastAndBeforeCursor(t *testing.T) {
 	require.NoError(t, err)
 
 	marketID := testData.markets[0].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{}, testData.orders[3], testData.orders[7], testData.orders[13])
@@ -1824,7 +1828,7 @@ func testOrdersCursorPaginationByPartyLastAndBeforeCursor(t *testing.T) {
 	require.NoError(t, err)
 
 	partyID := testData.parties[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{}, testData.orders[5], testData.orders[7], testData.orders[9])
@@ -1881,7 +1885,7 @@ func testOrdersCursorPaginationByMarketAndPartyLastAndBeforeCursor(t *testing.T)
 
 	partyID := testData.parties[1].ID.String()
 	marketID := testData.markets[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 1)
 	want := append([]entities.Order{}, testData.orders[9])
@@ -1909,7 +1913,7 @@ func testOrdersCursorPaginationByMarketLastAndBeforeCursorNewestFirst(t *testing
 	require.NoError(t, err)
 
 	marketID := testData.markets[0].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{},
@@ -1941,7 +1945,7 @@ func testOrdersCursorPaginationByPartyLastAndBeforeCursorNewestFirst(t *testing.
 	require.NoError(t, err)
 
 	partyID := testData.parties[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, nil, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 3)
 	want := append([]entities.Order{},
@@ -2006,7 +2010,7 @@ func testOrdersCursorPaginationByMarketAndPartyLastAndBeforeCursorNewestFirst(t 
 
 	partyID := testData.parties[1].ID.String()
 	marketID := testData.markets[1].ID.String()
-	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination)
+	got, pageInfo, err := stores.os.ListOrders(ctx, &partyID, &marketID, nil, false, pagination, entities.DateRange{})
 	require.NoError(t, err)
 	assert.Len(t, got, 1)
 	want := append([]entities.Order{},
@@ -2018,5 +2022,102 @@ func testOrdersCursorPaginationByMarketAndPartyLastAndBeforeCursorNewestFirst(t 
 		HasPreviousPage: true,
 		StartCursor:     testData.cursors[9].Encode(),
 		EndCursor:       testData.cursors[9].Encode(),
+	}, pageInfo)
+}
+
+func testOrdersCursorPaginationBetweenDatesByMarketNoCursor(t *testing.T) {
+	stores, teardown := setupOrderCursorPaginationTests(t)
+	defer teardown(t)
+	testData := generateTestOrdersForCursorPagination(t, stores)
+
+	t.Logf("Test DB Port: %d", testDBPort)
+
+	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
+	defer cancel()
+	pagination, err := entities.NewCursorPagination(nil, nil, nil, nil, false)
+	require.NoError(t, err)
+
+	marketID := testData.markets[0].ID.String()
+	startDate := testData.orders[3].VegaTime
+	endDate := testData.orders[14].VegaTime
+	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination, entities.DateRange{
+		Start: &startDate,
+		End:   &endDate,
+	})
+
+	require.NoError(t, err)
+	assert.Len(t, got, 2)
+	want := append([]entities.Order{}, testData.orders[3], testData.orders[7]) // order[13] and order[14] have the same block time
+	assert.Equal(t, want, got)
+	assert.Equal(t, entities.PageInfo{
+		HasNextPage:     false,
+		HasPreviousPage: false,
+		StartCursor:     testData.cursors[3].Encode(),
+		EndCursor:       testData.cursors[7].Encode(),
+	}, pageInfo)
+}
+
+func testOrdersCursorPaginationBetweenDatesByMarketFirstCursor(t *testing.T) {
+	stores, teardown := setupOrderCursorPaginationTests(t)
+	defer teardown(t)
+	testData := generateTestOrdersForCursorPagination(t, stores)
+
+	t.Logf("Test DB Port: %d", testDBPort)
+
+	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
+	defer cancel()
+	first := int32(3)
+	pagination, err := entities.NewCursorPagination(&first, nil, nil, nil, false)
+	require.NoError(t, err)
+
+	marketID := testData.markets[0].ID.String()
+	startDate := testData.orders[3].VegaTime
+	endDate := testData.orders[14].VegaTime
+	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination, entities.DateRange{
+		Start: &startDate,
+		End:   &endDate,
+	})
+	require.NoError(t, err)
+	assert.Len(t, got, 2)
+	want := append([]entities.Order{}, testData.orders[3], testData.orders[7])
+	assert.Equal(t, want, got)
+	assert.Equal(t, entities.PageInfo{
+		HasNextPage:     false,
+		HasPreviousPage: false,
+		StartCursor:     testData.cursors[3].Encode(),
+		EndCursor:       testData.cursors[7].Encode(),
+	}, pageInfo)
+}
+
+func testOrdersCursorPaginationBetweenDatesByMarketFirstAndAfterCursor(t *testing.T) {
+	stores, teardown := setupOrderCursorPaginationTests(t)
+	defer teardown(t)
+	testData := generateTestOrdersForCursorPagination(t, stores)
+
+	t.Logf("Test DB Port: %d", testDBPort)
+
+	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
+	defer cancel()
+	first := int32(3)
+	after := testData.cursors[0].Encode()
+	pagination, err := entities.NewCursorPagination(&first, &after, nil, nil, false)
+	require.NoError(t, err)
+
+	marketID := testData.markets[0].ID.String()
+	startDate := testData.orders[0].VegaTime
+	endDate := testData.orders[14].VegaTime
+	got, pageInfo, err := stores.os.ListOrders(ctx, nil, &marketID, nil, false, pagination, entities.DateRange{
+		Start: &startDate,
+		End:   &endDate,
+	})
+	require.NoError(t, err)
+	assert.Len(t, got, 2)
+	want := append([]entities.Order{}, testData.orders[3], testData.orders[7])
+	assert.Equal(t, want, got)
+	assert.Equal(t, entities.PageInfo{
+		HasNextPage:     false,
+		HasPreviousPage: true,
+		StartCursor:     testData.cursors[3].Encode(),
+		EndCursor:       testData.cursors[7].Encode(),
 	}, pageInfo)
 }