+++ datanode/sqlstore/stake_linking.go
@@ -36,8 +36,8 @@ foreign_tx_hash, log_index, ethereum_address, tx_hash, vega_time`
 )
 
 var stakeLinkingOrdering = TableOrdering{
-	ColumnOrdering{"vega_time", ASC},
-	ColumnOrdering{"id", ASC},
+	ColumnOrdering{Name: "vega_time", Sorting: ASC},
+	ColumnOrdering{Name: "id", Sorting: ASC},
 }
 
 func NewStakeLinking(connectionSource *ConnectionSource) *StakeLinking {